# Server/src/main/java/fr/eseo/twic/milokoson/dto/Order/OrderDto.java
package fr.eseo.twic.milokoson.dto.Order;

import fr.eseo.twic.milokoson.bo.Customer;
import fr.eseo.twic.milokoson.bo.OrderStatus;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.springframework.hateoas.RepresentationModel;

import java.time.LocalDateTime;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
public class OrderDto extends RepresentationModel<OrderDto> {

    private String orderId;
    private Customer customer;
    private OrderStatus orderStatus;
    private LocalDateTime placedTimestamp;
    private LocalDateTime deliveredTimestamp;
    private LocalDateTime cancelledTimestamp;

}