# Server/src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Rest Milo KOSON",
    "description": "API for managing customers, orders, order details, order statuses, and products",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1.0",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/customers": {
      "get": {
        "summary": "Get all customers",
        "responses": {
          "200": {
            "description": "List of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      }
    },
    "/search/{accountNo}": {
      "get": {
        "summary": "Search customers by account number",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers matching the account number",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customers/{accountNo}": {
      "get": {
        "summary": "Get details of a customer by account number",
        "parameters": [
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{orderId}": {
      "put": {
        "summary": "Update an existing order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted"
          }
        }
      }
    },
    "/orderDetails": {
      "get": {
        "summary": "Get all order details",
        "responses": {
          "200": {
            "description": "List of order details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDetailsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orderDetails/{orderId}": {
      "get": {
        "summary": "Get order details by order ID",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of order details for the specified order ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDetailsDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create order details for an order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OrderDetailsDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order details created"
          }
        }
      }
    },
    "/orderDetails/{orderId}/{orderDetailId}": {
      "put": {
        "summary": "Update an existing order detail",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderDetailId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated order detail"
          }
        }
      },
      "delete": {
        "summary": "Delete an existing order detail",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderDetailId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order detail deleted"
          }
        }
      }
    },
    "/orders/orderStatuses": {
      "get": {
        "summary": "Get all order statuses",
        "responses": {
          "200": {
            "description": "List of order statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderStatusDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order status",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created order status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatusDto"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "accountNo": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "registrationTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "orderStatus": {
            "$ref": "#/components/schemas/OrderStatusDto"
          },
          "placedTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "deliveredTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledTimestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderDetailsDto": {
        "type": "object",
        "properties": {
          "orderDetailId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "productId": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          }
        }
      },
      "OrderStatusDto": {
        "type": "object",
        "properties": {
          "orderStatusId": {
            "type": "string"
          },
          "orderStatusName": {
            "type": "string"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productNo": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      }
    }
  }
}